+++ src/main/java/store/mybooks/front/utils/Utils.java
@@ -1,6 +1,7 @@
 package store.mybooks.front.utils;
 
 import java.util.List;
+import java.util.Objects;
 import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -75,11 +76,15 @@ public class Utils {
         return headers;
     }
 
-    public static HttpHeaders addAuthHeader(HttpServletRequest request) { // todo addHeader 이름변경하기
 
+    public static HttpHeaders addAuthHeader(HttpServletRequest request) {
         HttpHeaders headers = Utils.getHttpHeader();
-        String token = Utils.getIdentityCookieValue(request.getCookies());
-        headers.set("Authorization", token);
+
+        String token = (String) request.getAttribute("identity_cookie_value");
+        if (Objects.isNull(token)) {
+            throw new AuthenticationIsNotValidException();
+        }
+        headers.set("Authorization", (String) request.getAttribute("identity_cookie_value"));
         return headers;
     }
 
@@ -103,17 +108,7 @@ public class Utils {
         return exchange.getBody();
     }
 
-    public static String getIdentityCookieValue(Cookie[] cookies) {
 
-        if (cookies != null) {
-            for (Cookie cookie : cookies) {
-                if ("identity_cookie".equals(cookie.getName())) {
-                    return cookie.getValue();
-                }
-            }
-        }
-        throw new AuthenticationIsNotValidException();
-    }
 
 
     public static HttpHeaders getAuthHeader() {
@@ -121,29 +116,7 @@ public class Utils {
                 .getAttribute("authHeader", RequestAttributes.SCOPE_REQUEST);
     }
 
-    public static void addJwtCookie(HttpServletResponse response, String token) {
 
-        response.setHeader("Set-Cookie",
-                "identity_cookie=" + token + "; " +
-                        "Path=/; " + // 적용될 범위
-                        "Domain=localhost; " + // 적용될 도메인
-                        "HttpOnly; " + // JavaScript에서 쿠키에 접근하는 것을 방지하기 위해 HttpOnly 속성을 설정합니다.
-                        "Max-Age=604800; " + // 쿠키 생존시간
-                        "SameSite=Strict; " +
-                        // SameSite 설정 (Strict, Lax, None 중 선택) Strict 쿠키가 같은 도메인에서만 헤더로 넘어감, 요청보낼떄는 헤더에 담아 보낼꺼임
-                        "Secure" // Secure 설정
-        );
-    }
-
-    public static void deleteJwtCookie(HttpServletResponse response) {
-        response.setHeader("Set-Cookie",
-                "identity_cookie=; " +
-                        "Path=/; " +
-                        "Domain=localhost; " +
-                        "Max-Age=0; " + // 쿠키를 즉시 만료시킵니다.
-                        "SameSite=Strict; " +
-                        "Secure");
-    }
 
 
 }
