+++ src/main/java/store/mybooks/front/utils/interceptor/CookieInterceptor.java
@@ -3,6 +3,7 @@ package store.mybooks.front.utils.interceptor;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import org.springframework.web.servlet.HandlerInterceptor;
+import store.mybooks.front.utils.CookieUtils;
 import store.mybooks.front.utils.Utils;
 
 /**
@@ -25,7 +26,7 @@ public class CookieInterceptor implements HandlerInterceptor {
         System.out.println("요청 URI: " + request.getRequestURI());
         // 컨트롤러 부를때 한번 부름 ,  /user/asserts/images/logo/white-logo .jpg .svg
         // 얘들 이름 바꿔주면 됨
-        request.setAttribute("identity_cookie_value", Utils.getIdentityCookieValue(request));
+        request.setAttribute("identity_cookie_value", CookieUtils.getIdentityCookieValue(request));
 
         return true; // 요청 계속 진행
     }
